requirements.txt
streamlit
Pillow
ollama
opencv-python
numpy
fpdf
torch
torchvision
segmentation-models-pytorch